feat(pluggable-widgets-mcp): refactor security guardrails into a single ts, update docs
